Moved private protobuf object constructors
Replaced “protobuf_utils” with separate *_priv.h headers for each object
