+++ src/crypto/root_keys.c
@@ -1,7 +1,9 @@
 
 #include "root_keys.h"
 #include "memory.h"
-#include "protobuf_util.h"
+#include "storage.pb-c.h"
+#include "cipher_priv.h"
+#include "eckey_priv.h"
 
 #define CURRENT_ROOT_KEY_VERSION 1
